refactor(src): split code into files
